New rule for file level retries (#4120)
Much like the unittest flaky rule, this isn't a rule that points to a
line that explains why the job failed, but makes it easier to find jobs
that needed to do retries. It's relatively near the bottom of the rules
to make sure it doesn't interfere with the rules that do point to
failing lines.