Adding option to run without docker to linux generic job (#4211)
TorchRL requires some scripts to be run outside of the docker container.
Previously, the generic linux_job requires all jobs to be run inside a
conda-builder container. This PR introduces the option for scripts to be
run on bare metal, but still fetches the docker container in case other
scripts need to be run on the container.