Don't check for authorization in issue_query (#4208)
I originally considered adding pytorchbot as a special exception, but I
think this is fine because issues are almost immediately closed if they
don't have authorization.

Also I'm a little worried that reopening an old issue that was created
by someone without write permissions would not disable the test. While
this is acceptable since a new issue could also be created, I think it
creates some confusion between which issues work and which don't. Also
not sure what happens to the author association in rockset when people
leave/join pytorch.