Alert for long machine queues (#4180)
Create an alert for long queues

The limit is 4 hours or 75 machines for most machines and a separate
category for the gcp a100 runners since they generally have longer queue
times but fewer machines. The avg non zero queue seems to be about 2.5
hours and 40 machines for most machine types.

It uses a new label queue-alerts and attempts to re open issues when
there is an issue with the queue-alert label if it is closed.