Fix revert printer (#4041)
Attempt to make revert printer slightly more robust against weird
comments by re-parsing the comment. This is probably a bad idea

* also change mention on master to main
* change sql query to be more general, also put sql query directly into
the script so its easier to change
* sort the categories 
* added comment url to the revert body in trymerge so use that to
determine which comment caused the revert. This also means you can get
rid of the "comment but no corresponding commit" section, but it also
means that it this script won't work on commits before the change to
trymerge

test plan:
ran it against the past three weeks and compared against old versions. I
actually found that the old printer was unable to find some comments and
incorrectly put the revert into the "not through pytorchbot section"