default the linux_job workflow to pull submodules recursively
Summary:
When trying to use this for the "codegen diff" job in PyTorch, I
noticed it was failing because we did not have submodules
available. After discussing with @seemethere, he suggested to just
make this the default.

Test Plan:
Verifying in local CI.

This can't be tested in the pytorch/pytorch repo because I this is
pushed to the dagitses/test-infra fork, which fails with the error:
> Called workflows cannot be queued onto self-hosted runners across
> organisations/enterprises. Failed to queue this job.
> Labels: 'linux.large'.

See https://github.com/pytorch/pytorch/actions/runs/5169464586/attempts/2