Trigger workflows on HUD (#4303)
This PR adds the option to quickly trigger a workflow on a HUD commit,
which is useful to debug:

* Periodic failures
* Breaking ghstack commit in the middle of the stack

By clicking a checkbox, a logged in user could now do the following step
on HUD to start a workflow:

```
git checkout SHA
git tag ciflow/WORKFLOW_NAME/SHA-DEBUG
git push origin ciflow/WORKFLOW_NAME/SHA-DEBUG
```

Here are some details:
* This only supports for trunk, periodic, and slow workflows on PyTorch
atm (may be I should add binaries if there is a need for that)
* The list of workflows is at the bottom of the commit page after all
the workflow jobs. This is to avoid people accidentally clicking on it
and start these workflows unnecessarily
* If the workflow has already been triggered. The list will be empty
* The list is only on HUD commit page atm as the PR already has `ciflow`
label mechanism
* Only logged in users can see the list
* Once logged in, only people with write or admin access to the repo can
start the workflow. This is the same requirements for using `ciflow`
label in PR
* The tag is in the format `ciflow/WORKFLOW_NAME/SHA`

### Testing

* Not logged in, nothing is shown
https://torchci-git-fork-huydhn-add-periodic-flow-tag-fbopensource.vercel.app/pytorch/pytorch/commit/1b16ac7481628b2a2fe96b43056573a7ee3c2387
* Logged in, local dev

![Screenshot 2023-06-17 at 14 14
06](https://github.com/pytorch/test-infra/assets/475357/8aa06d37-8409-4313-bfa7-917129829ff2)