add scripts to generate github issues from alerts
add scripts to generate github issues from alerts

add scripts to generate github issues from alerts.

In this pr we create a registry to associate functions to publish alerts with the alerts themselves. We enable recurrently failing alerts to do this. We test by using a github action which has a workflow-dispatch.

Example Issue: #4267

copilot:all