Remove special cases for torchaudio FFmpeg
https://github.com/pytorch/audio/pull/3460 changed the way torchaudio is built,
and there is no longer need to supply FFmepg externally.
test-infra no longer needs to build FFmpeg and use pre/post-scripts for torchaudio.

The only issue remaining for now is smoke test.
Currently torchaudio's smoke test script by default expects that FFmpeg libraries
are available. This can be disabled with `--no-ffmpeg`, but the way YML files are
written in test-infra repo does not allow to pass a flag to smoke test.

We can switch the behavior, or use other smoke test.