Remove special cases for torchaudio FFmpeg (#4358)
This commit removes all the code related to FFmpeg, which are specific
to torchaudio.

https://github.com/pytorch/audio/pull/3460 changed the way torchaudio is
built. FFmpeg integration is handled automatically and enabled by
default. This eliminates the need to supply FFmepg externally.
test-infra no longer needs to build FFmpeg and use pre/post-scripts for
torchaudio.

Note that smoke test was updated in
https://github.com/pytorch/audio/pull/3465 to not use FFmpeg for wheel.
We will revisit this later with multiple FFmpeg version support.