Add Unittests for Build Matrix Generation (#4386)
Adds Unittests for build matrix generation. This PR does a couple
things:
* Refactors the build matrix generation script to have a general-purpose
function that returns the matrix as a python dict
* Adds JSON files in `tools/tests/assets` that contain the output of
running the generation script with common configs (not comprehensive atm
but covers the most common 6 configurations - Wheels/Conda across the 3
OS's with CUDA enabled)
* Adds Unittests that call the generic matrix generation function and
compares it with the serialized matrices in the JSON files.

The unittests will be picked up by the GHA job in `tests.yml`.