Fix the wrong use of Map with Object.keys in getBaseCommitJobs (#4391)
[Object.keys](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/keys)
iterates though all string-keyed property names in the object. However,
this works with `{}` but not with a `Map` object. In fact,
`Object.keys(mapObject)` returns an empty array.

Getting the empty list of base commit jobs means Dr.CI fails to detect
broken trunk failures, i.e.
https://github.com/pytorch/pytorch/pull/105145

### Testing

This is a big gotcha moment for me, so I add an unit test for
`getBaseCommitJobs` to properly test this