infra: Switch nvidia template to utilize regular cloudwatch config
We were running into issues where the cloudwatch-agent-ctl was
attempting to call out to nvidia-smi, but we actually install that as a
later part of our workflow. Since we don't currently use the GPU
statistics that we store in cloudwatch this is a temporary fix to get
GPU instances unblocked.

Error experienced:

```
E! [telegraf] Error running agent: validate input plugin nvidia_smi failed because of Cannot get file's stat /usr/bin/nvidia-smi: no such file or directory
```

NOTE: This is a short term fix and should be treated as such, the longer
term fix would be to have an AMI that has nvidia-smi pre-installed but
that is a lot to ask for in an afternoon

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>