Add dropdown to hud header (#4446)
Referenced css from
https://blog.logrocket.com/how-create-multilevel-dropdown-menu-react/
and its source code at
https://github.com/Ibaslogic/react-multilevel-dropdown-menu

On my smaller screen, the HUD header was becoming a bit full, so I added
a dropdown and moved a lot of the dev infra links into it. I left
metrics in the bar since I think it is used frequently, but I am open to
moving links around if people want fewer clicks for pages they visit
often.

The dropdown component is reusable so if we also want to add a lot of
repositories and the header doesn't have space for that, we can move
them all into a dropdown as well. I have no idea how the dropdown reacts
to having so many things in it it takes up the entire page but that's a
problem for future me (probably either make it tiered or scrollable).

Also remove the main branch link since the homepage is main.

Please visit the vercel preview to see if you like it.

Fixes https://github.com/pytorch/test-infra/issues/4439