Query and match failure line when detecting flaky job (#4449)
When identifying flaky job using regex rule, it's easier to use the
whole failure line instead of just the captured string. Here is the case
that baffled me yesterday
https://hud.pytorch.org/pr/pytorch/pytorch/106270.

1. `macos-12-py3-arm64 / test (default, 2, 3, macos-m1-12)` failed
flakily (SIGSEGV) and the error line was `RuntimeError:
inductor/test_torchinductor_opinfo 2/2 failed! Received signal: SIGSEGV`
1. However, the capture string was only `RuntimeError:
inductor/test_torchinductor_opinfo 2/2 failed!`, not the whole line.

Even if I want to write a regex rule to capture the crash (SIGSEGV), I
couldn't do so because it wasn't captured. Also it would be confusing
when writing regex rule for captured string because we only see the
failure line on HUD.

I also update the 2 Rockset query with comments to clearly say how they
are used.