Remove no-op force merges masquerading as impatient force merges (#4486)
This is a proposal to remove no-op force merges masquerading as
impatient force merges. IMO, a legit force merge needs to satisfy one of
the two conditions below:

1. `skip_mandatory_checks` is true (-f) and `failed_checks_count > 0`
(with failures) or `pending_checks_count > 0` (impatience). Under this
condition, If a force merge (-f) is done when there is no failure and
all jobs have finished, it's arguably just a regular merge in disguise.
This is the main point of this proposal. Here is an example of a no-op
force merge
https://github.com/pytorch/pytorch/pull/106912#issuecomment-1680999969
(out of many)
2. `ignore_current` is true (-i) and `is_failed` is false (indicating a
successful merge) and `ignored_checks_count > 0` (with failures). As
`-i` still waits for all remaining jobs to finish, this shouldn't be
counted toward force merge due to impatience

If none applies, the merge should be counted as a regular merge
regardless of the use of `-f` or `-i`. We could track that (regular
merges masquerading as force merges) to understand how devs use (or
abuse) these flags, but that should be tracked separately IMO because
it's arguably a different case altogether.

Technically, this uses `pending_checks` field from the `merges`
collection that we haven't made use of till date to remove a significant
portion of no-op force merges masquerading as impatient force merges.
IMO, impatient force merges is not the negative part of force merge with
failures, thus the way these queries are atm is a bit wrong (a force
merge is either due to failures or due to impatience, there is no
other). This PR fixes them according to the definition above.

### Testing


https://torchci-git-fork-huydhn-force-merge-no-pending-fbopensource.vercel.app/metrics.
As expected, % force merges with failures remains unchanged but we get a
more accurate view of % impatient force merges. The latter now requires
`pending_checks_count > 0`.

Same goes for
https://torchci-git-fork-huydhn-force-merge-no-pending-fbopensource.vercel.app/kpis.