Skip linting nonexitent files (#57)
* Allow multiple toml files in config


This pr allows the usage of multiple configuration files with lintrunner. If a conflict occurs the latter file in the order they are entered takes precedence.

This allows us to have custom actions for lintrunner internally, and for users of lintrunner personal configurations if they have weird branch names or something of that manner.

* Skip linting nonexitent files

Sometimes sapling / git picks up ephemeral files. This pr allows

---------

Co-authored-by: PaliC <>