Merge branch 'main' into pr60 (#60)
* Allow multiple toml files in config


This pr allows the usage of multiple configuration files with lintrunner. If a conflict occurs the latter file in the order they are entered takes precedence.

This allows us to have custom actions for lintrunner internally, and for users of lintrunner personal configurations if they have weird branch names or something of that manner.

* Skip linting nonexitent files

Sometimes sapling / git picks up ephemeral files. This pr allows

* Add command line argument `--only-lint-under-config-dir`

Adds command line argument `--only-lint-under-config-dir` option to only lint under the configuration directory. For simplicty this argument can also be added to a configuration file as well.

---------

Co-authored-by: PaliC <>