Maintain old test times (#4509)
In https://github.com/pytorch/test-infra/pull/4476 I removed the code
that allows us to edit the test times manually and have it mostly
persist because I thought we didn't need it anymore but I was wrong.
This PR adds this ability back in.

This means the test times json never shrinks, but I think that's ok
since we can manually edit it.

Also
* rename the job
* move to ubuntu-latest