[dr ci] Fix too many jobs getting removed by `removeCancelledJobAfterRetry` (#4610)
The `removeCancelledJobAfterRetry` #4579 was not taking into account
that there are jobs coming from multiple commits, so it ended up finding
jobs from other commits/PRs and thinking that it was a retry/the same
job so it got removed.

My solution is to call it per PR, but it could probably be made to take
into account commit sha or something

Other changes:
* This PR pretty much reverts #4604 in favor of querying for the full
name and the job name, since I realized jobs that exist both on trunk
and pull (windows, rocm builds) would get removed by
`removeCancelledJobAfterRetry` if the workflow name wasn't included in
the name. There are other solutions to this, but I figured this would be
ok since we don't actually use the workflow name anywhere else. This
also means that the windows/rocm builds that exist on both pull and
trunk will be seen as different jobs by the classifier, so broken trunk
on pull will not register for trunk. I don't know how prevalent this is,
but now that we have both jobname and name it should be easy to change.
* Makes job = full name with workflow name, and jobName = the job's name
= what Dr CI used to call name which is how most of the other queries do
it.
* Jobs in the pr info is now an array since the key was never actually
used outside of deduping retries
* Made `removeCancelledJobAfterRetry` not async

I think this fixes #4598

Tested on 109584 in pytorch