Add an excluded list for flaky detection on Dr.CI (#4636)
https://github.com/pytorch/pytorch/pull/111207 was merged because Dr.CI
detects the lint failure as flaky. This was an FP because it found an
exact match of the same lint failure, on a different branch. This
happens because lint failure only prints the name of the broken file.

```
{
    name: 'Lint / lintrunner / linux-job',
    workflowName: 'Lint',
    jobName: 'lintrunner / linux-job',
    sha: '2b30b63b86ee277b3e70d39db6244ac0360c84c3',
    id: 17653166474,
    workflowId: 6499566593,
    time: '2023-10-12T18:59:16Z',
    conclusion: 'failure',
    htmlUrl: 'https://github.com/pytorch/pytorch/actions/runs/6499566593/job/17653166474',
    failureLine: '>>> Lint for torch/_dynamo/output_graph.py:',
    failureLineNumber: 145946,
    failureCaptures: [ '>>> Lint for torch/_dynamo/output_graph.py:' ]
}
```

In hindsight, lint jobs are stable enough that we shouldn't need to try
to see if they are flaky.