Add the list of previous commands as optional context on log classifier (#4649)
I have this change to add the list of N (5) previous commands as
optional context on log classifier. This relies on GitHub console log
convention to prefix a **bash** command with `+`. This new `context`
field should just become a new field on `workflow_job` for future query.

This is mainly an attempt to provide more context on GHA generic
failures that are not well captured by the log classifier.