Fix the number of hooks conditionally rendered by hasArtifacts (#4677)
Keep the number of hooks the same but hide the <a> tag if there is no
artifacts. This probably fixes the error reported on Tree Hugger when
loading https://hud.pytorch.org/pr/111935

Also we cannot use hook, i.e. `useState` in `WorkflowBox` because the
number of jobs can change dynamically on PR or commit pages.