Add isSameAuthor check when classifying flaky tests with Dr.CI (#4688)
I strengthen the `isSameHeadBranch` check to become `isSameAuthor` to
avoid an FP in flaky detection when the same author submits multiple PRs
of the same change. For example,

* Cherry picking https://github.com/pytorch/pytorch/pull/111845
* Draft PR and then submit a real PR (I remember seeing an example of
this, but couldn't find the PR number now)

I choose to use author email here because that's readily available in
the commit info of the workflow run. The commit info, however, is not
available in `torch-workflow-job` DynamoDB table which powers the
search. So, an additional Rockset query is needed foreach flaky match.

### Testing

```
curl --request POST \
    --url 'http://localhost:3000/api/drci/drci?prNumber=111845' \
    --data 'repo=pytorch'
```