Handle null machine_type for queued_jobs queries (#4788)
Fixes https://github.com/pytorch/test-infra/issues/4784. This handles
the edge case where there is no label associated with the runner, thus
the `machine_type` field is empty. This error happens today with some
ROCm runners, probably due to the current ROCm outage. AMD is cleaning
this up, but we can have this change to harden the HUD page a bit.