Specific PYTORCH_VERSION when building RC for test channel (#4804)
Cherry pick https://github.com/pytorch/test-infra/pull/4803 for 2.1.2
release. This is to fix the issue when building 2.1.2 RC for domains
pick up 2.2.0 from test channel instead of 2.1.2

The changes in `tools/tests/assets` are not really needed but I like a
green signal for `python3 -m unittest discover -vs tools/tests -p
'test_*.py'`. The differences are in the pinned Docker images.