Specific PYTORCH_VERSION when building RC for test channel (#4803)
With 2 candidate RC in the test channel (2.1.2 and 2.2.0), we need to
have a way to specify which PYTORCH_VERSION to use. The issue surfaces
when we are trying to cut RC3 for 2.1.2 today while there is already RC1
for 2.2.0 in place. In the current state, `PYTHON_VERSION=3.11
PACKAGE_TYPE=conda CU_VERSION=cpu ARCH_NAME=arm64 CHANNEL=test python -m
pytorch_pkg_helpers` will set `PYTORCH_VERSION` to 2.2.0 whenever
domains are built

For example,
https://github.com/pytorch/vision/actions/runs/7190086502/job/19582646141#step:6:743
build M1 conda for vision 0.16.2 and it's expected to use 2.1.2 there,
not 2.2.0.

The fix here is to:

1. Correctly set the channel to `test` when building RC
2. If the channel is test, use the current candidate version as
PYTORCH_VERSION. The value is exposed via the build matrix as
`stable_version`
3. When doing branch cut for test-infra, the `CURRENT_CANDIDATE_VERSION`
variable could be set accordingly, for example, use 2.1.2 for
`release/2.1` and 2.2.0 for `release/2.2`

### Testing plan

This PR https://github.com/pytorch/test-infra/pull/4803 tests `nightly`
while the cherry pick https://github.com/pytorch/test-infra/pull/4804
covers `test` channel