Migrate update-commit-hash GHA to test-infra (#4885)
So that it can be re-used by ExecuTorch to update PyTorch pinned
commits. This is a close 1:1 copy of the GHA and the script from
`pytorch/pytorch`. A small modification is made in the `source-repo`
parameter to allow the pinned commit files from other repo, i.e.
`pytorch/executorch`.

### Testing

Migrate and test the new GHA on pytorch
https://github.com/pytorch/pytorch/pull/117506:

* ExecuTorch PR already exists
https://github.com/pytorch/pytorch/actions/runs/7534593538/job/20509371346#step:2:1114
* Audio has no new commit
https://github.com/pytorch/pytorch/actions/runs/7534593538/job/20509371162#step:2:1107
* Vision has a new commit and a PR should be created, but GH correctly
reject `git push` to create a new remote branch
https://github.com/pytorch/pytorch/actions/runs/7534593538/job/20509371569#step:2:1109.
This should work when running on nightly schedule.

Also testing this out on ExecuTorch
https://github.com/pytorch/executorch/actions/runs/7535757395/job/20513578984?pr=1597#step:2:1255