Fix the wrong upload of vs2019 conda package on Windows (#4893)
In the new setup, the conda upload job will just upload any `tar.bz2`
packages it finds because it doesn't have any additional information
about the target. This leads to the wrong and duplicated upload of
vs2019 conda package in Windows build job, i.e.
https://github.com/pytorch/vision/actions/runs/7555102811/job/20584135167#step:7:170.

I'm not sure what's the best way to fix this here, but removing this
file explicitly helps. Also, this only happens on Windows.