Resurrect the option to force push the Docker image (#4866)
IIRC, we used to have this option, but nothing was using it. I have
found an use case for this in
https://github.com/pytorch/executorch/pull/1335#discussion_r1448247422
where ExecuTorch team copies the content of `examples/arm` into
`.ci/docker` before starting the build
https://github.com/pytorch/executorch/blob/main/.ci/docker/build.sh#L40-L46:

* The docker image was rebuilt, but its hash tag remained the same
because the copy step above is not version tracked.
* The new image was not pushed.

I have tried to move `examples/arm` into `.ci/docker` and setup
softlinks, but the result looks weird
https://github.com/pytorch/executorch/pull/1574#discussion_r1447762323,
so I'll probably drop that approach.

The remaining option here is to add a `force-push` flag that can be used
to, well, force push the image even if its hash tag already exists.