Copyable Reproduction Fix (#4632)
This should fix the less-helpful reproduction we were producing when the
"copy repro" button was pressed on the HUD. This hasn't been tested as I
don't have access to the Vercel deployment.

I will also begin working on the bash script discussed earlier, and will
likely be at office hours today to discuss it.

cc @huydhn

---------

Co-authored-by: Huy Do <huydhn@gmail.com>