[ez] uploadQueryLambdas - Allow description for lambdas to be different types (#5039)
Sometimes the description from the query lambda on rockset is null and
the local description is empty, and the script attempts to update the
lambda. This made me hit rate limits when trying to update from local.
I assume this is just a typing thing and that this is just two different
ways of saying that there is no description, so add a check for that.