Support the upload of custom artifacts from artifacts-to-be-uploaded folder (#5092)
This is to support a new use case where I need to upload different
Android and iOS apps instead of the regular wheel files as the artifacts
of Nova Linux job. It looks like the easiest way is to have a folder
`artifacts-to-be-uploaded` in the current workspace, and let the job
script moves any artifacts it needs to persist there.