Add Android support to mobile job workflow (#5077)
This extends the current mobile job workflow to support instrumentation
tests on Android
https://docs.aws.amazon.com/devicefarm/latest/developerguide/test-types-android-instrumentation.html.
As part of this change, I also add the support to upload an extra data
archive to the device to be used by the test. In the context of
performance benchmark, the extra data archive usually contains exported
models that the benchmark needs to cover.

### Testing

I'm using ExecuTorch llama2 demo app as inputs of the test
https://github.com/pytorch/executorch/tree/main/examples/demo-apps/android/LlamaDemo.
It uses the exported llama2 7B model.

The script can be run locally as follows and its inputs can be
downloaded from S3 `ossci-assets` bucket:

```
python run_on_aws_devicefarm.py \
    --project-arn "arn:aws:devicefarm:us-west-2:308535385114:project:02a2cf0f-6d9b-45ee-ba1a-a086587469e6" \
    --device-pool-arn "arn:aws:devicefarm:us-west-2:308535385114:devicepool:02a2cf0f-6d9b-45ee-ba1a-a086587469e6/e59f866a-30aa-4aa1-87b7-4510e5820dfa" \
    --app ~/executorch/examples/demo-apps/android/LlamaDemo/app/build/outputs/apk/debug/app-debug.apk \
    --android-instrumention-test ~/executorch/examples/demo-apps/android/LlamaDemo/app/build/outputs/apk/androidTest/debug/app-debug-androidTest.apk \
    --extra-data "arn:aws:devicefarm:us-west-2:308535385114:upload:02a2cf0f-6d9b-45ee-ba1a-a086587469e6/bd15825b-ddab-4e47-9fef-a9c8935778dd" \
    --test-spec ~/android-llama2-device-farm-test-spec.yml \
    --name-prefix "DEBUG-ANDROID-LLAMA2" \
    --workflow-id "1" \
    --workflow-attempt "1"
```

Because these inputs are already on AWS, the script can also refer to
them directly without re-uploading. This is needed because the exported
llama2 7b model is 3+ GB.

```
python run_on_aws_devicefarm.py \
    --project-arn "arn:aws:devicefarm:us-west-2:308535385114:project:02a2cf0f-6d9b-45ee-ba1a-a086587469e6" \
    --device-pool-arn "arn:aws:devicefarm:us-west-2:308535385114:devicepool:02a2cf0f-6d9b-45ee-ba1a-a086587469e6/e59f866a-30aa-4aa1-87b7-4510e5820dfa" \
    --app "arn:aws:devicefarm:us-west-2:308535385114:upload:02a2cf0f-6d9b-45ee-ba1a-a086587469e6/56b109f3-64b4-48c0-b1c8-b19dafc80647"  \
    --android-instrumentation-test "arn:aws:devicefarm:us-west-2:308535385114:upload:02a2cf0f-6d9b-45ee-ba1a-a086587469e6/528113d6-2003-420b-aedb-4242b6d14254" \
    --extra-data "arn:aws:devicefarm:us-west-2:308535385114:upload:02a2cf0f-6d9b-45ee-ba1a-a086587469e6/bd15825b-ddab-4e47-9fef-a9c8935778dd" \
    --test-spec "arn:aws:devicefarm:us-west-2:308535385114:upload:02a2cf0f-6d9b-45ee-ba1a-a086587469e6/414cb54d-4d83-4576-8317-93244e4dc50e" \
    --name-prefix "DEBUG-ANDROID-LLAMA2" \
    --workflow-id "1" \
    --workflow-attempt "1"
```

The observed TPS is print to console, i.e.
https://github.com/pytorch/test-infra/actions/runs/8641760330/job/23692276850#step:12:70.
As this is run as part of JUnit test, I surface the value via an assert
statement for now and plan to improve that later.


cc @kirklandsign