Update ``update_dependencies.py`` script (#5109)
Users should now be able to specify a ``package`` for which the
dependencies should be updated. This also adds a flag to only check for
packages on PyPI (relevant when adding a new dependency). In addition,
the script now filters for the latest stable versions of packages to
avoid any weird beta version changes.