Exclude lint and backwards_compat from Dr.CI flaky check (#5106)
This is to address some new comments on
https://github.com/pytorch/test-infra/issues/5063.

1. Exclude `backwards_compat` from Dr.CI flaky check, as this job is
deem stable enough most of the time.
2. Lint jobs have already been excluded, but there is a bug when it
would still be treated as flaky when there is no associated log on S3.
This case surfaces in https://github.com/pytorch/pytorch/pull/124321
where there is a mix of signals from `pytorch` and` pytorch-canary` for
the same commit. For example, there is no
https://ossci-raw-job-status.s3.amazonaws.com/log/pytorch/pytorch-canary/24002161738
for the lint job there

```
{
  workflowId: 8745622961,
  workflowUniqueId: 13175283,
  id: 24002161738,
  runnerName: 'i-0cdd9180550988c58',
  authorEmail: 'ZainR@meta.com',
  name: 'Lint / workflow-checks / linux-job',
  jobName: 'workflow-checks / linux-job',
  conclusion: 'failure',
  completed_at: '2024-04-18T23:42:46Z',
  html_url: 'https://github.com/pytorch/pytorch-canary/actions/runs/8745622961/job/24002161738',
  head_branch: 'zainr/arn-fix',
  pr_number: 124321,
  head_sha: 'fa03725fa8512af211691a5164073ab7e2c4ee10',
  failure_captures: null,
  failure_lines: null,
  failure_context: null,
  time: '2024-04-18T23:42:51.131823Z'
}
```

The latter is probably a canary testing one-off thing, so I didn't
follow up further. But it's good to exclude lint properly anyway.