Support marking unstable jobs from other repos (#5122)
Fixes https://github.com/pytorch/executorch/issues/1759

Before this, a job can only be seen as unstable when it has the unstable
keyword in its name. The trick of renaming the job can be done with
PyTorch test jobs by manipulating the test matrix. However, that doesn't
work for PyTorch build job or Nova workflow without moving the filter
script to a separate job to run before them. That approach doesn't
scale.

So, this PR implements a different approach to query the list of open
unstable issues from Rockset from HUD, then group all jobs with an open
unstable issues to the unstable group.

The PR is big because the following top level pages need to query the
list and pass it around to different React components. The main update
is in the function `isUnstableJob` which calls `hasOpenUnstableIssue`.
The important call paths are as follows:

* HUD main page → `getGroupingData` → `classifyGroup` →
`hasOpenUnstableIssue`
* HUD commit/PR pages → `CommitStatus` → `FilteredJobList` →
`FailedJobInfo` → `JobSummary` → `isUnstableJob`
* HUD commit/PR pages → `CommitStatus` → `WorkflowsContainer` ->
`WorkflowBox` -> `WorkflowJobSummary` -> `JobSummary` → `isUnstableJob`

There are some misc pages like HUD flaky test pages that also needs to
be updated, but they don't need to have the list of unstable issues, so
I will just pass an empty array there.

The second part of this change is on Dr.CI.

### Testing

`trunk / test-coreml-delegate / macos-job` from ET is marked as unstable
by https://github.com/pytorch/executorch/issues/3264, and it seems to
show up correctly
https://torchci-git-support-unstable-job-other-repos-fbopensource.vercel.app/hud/pytorch/executorch/main.

An UX caveat is that all the invocations of the job will appear in the
unstable group because there is no `unstable` keyword to distinguish
them.