[HUD] Disable issue button - specify if issue is recent (#5166)
Maybe controversial change?  


Change the look of the disable/unstable button that links to the issue
that disables the test/marks the job unstable if the issue was not
updated recently. Changes it to the style that makes it seem like there
is no issue, so you have to click the button to make a new one, but the
button will still go to the old issue so you can reopen it.


I keep getting confused since when I see "previously disabled test" I
assume that the issue was wrongly closed recently and that I should
reopen, but then I discover that the issue is from 3 months ago and that
this is a new problem

Results in a change in expectations but it's not like "mark unstable
job" or "disable test" imply that a new issue always gets made.

I guess as a side note, if we don't do something like this, probably
every job is going to get the "previously unstable job" button
eventually

Example
old
<img width="611" alt="image"
src="https://github.com/pytorch/test-infra/assets/44682903/b84097b3-0c96-48e9-895e-9f856d1f1d07">

new

![image](https://github.com/pytorch/test-infra/assets/44682903/bedd3d7d-684a-4fb1-9cec-920727f59f73)