[EZ] [HUD] Always show a "Prev" button
UX improvement for when you want to scroll back (or "Next") multiple pages.

Previously, when you went to hud, scolled to the bottom, click the "Next" button, and scrolled down again, the "Next" button would have moved to the right because "Prev" suddenly appeared.

You no longer have to do that