Add benchmark dashboard for gpt-fast (#5268)
This is the initial version of the new LLMs benchmark dashboard with
data from gpt-fast. I reuse a bunch of components from the existing
inductor dashboard. One drawback here is that adding a new metric would
still require submitting changes to the dashboard, which is not ideal,
but would probably not a blocker neither. I'm pushing the initial
version of the dashboard first for early feedbacks from @yanboliang and
others.

### Testing


https://torchci-git-fork-huydhn-add-gpt-fast-dashboard-fbopensource.vercel.app/benchmark/llms
for look and feel. The data is coming from
https://github.com/pytorch/pytorch/blob/main/.github/workflows/inductor-micro-benchmark.yml
run on PyTorch CI